+++ common/views/vtkCrosshairRepresentation.h
@@ -60,8 +60,8 @@ public:
   vtkGetMacro(Opacity, double);
 //   vtkSetMacro(Opacity, double);
 
-  vtkSetVector3Macro(Position, int);
-  vtkGetVector3Macro(Position, int);
+  vtkSetVector3Macro(Position, double);
+  vtkGetVector3Macro(Position, double);
 
   //BTX
   vtkProp3D *GetCrosshairProp();
@@ -102,7 +102,7 @@ protected:
 
 protected:
   double Color;
-  int    Position[3];
+  double    Position[3];
   double Opacity;
   //ETX
 };
